#58 Adds diplomacyView start when AI noticed border crossing.
Same thing AI handles border crossing with AI. Only BUT here is
that handling is same as for regular diplomatic meeting. So it
is not really handling the border crossing. Now calls are ready for
start implementing border crossing handling.

Noticed also couple of bugs in AITurnView:
If last AI player noticed border crossing or new planet
diplomacy view was never really shown. Luckly this scenario
happened to be in testGame.save.