#89 Adds new Panel SpaceGreyPanel.
This should be used if background is not really
need to redrawn. There was a drawing issue with
invisible panel that combobox and rigid area are drawn with white
when selection menu is open. Fix was to add new panel to match
InfoPanel background color. ResearchView did not redraw
infoText now that invisible panel did not force to redrawing.