#124 Fixes duplicate fleet names.
I think this was because when creating scout fleets it only
calculate number fleet begining with name scout and increase the
number with 1. Problem was that AI was also attacking and it had
taken one scout ship into attack so number of fleet begining with
scout was less than actual last fleet name.

Removes also TODO to fix NPE since it was fixed earlier commit.