#190 Fixes numerous AI movement bugs.
Oh the horror. I found AI movement bugs. Let's go them through:
1. Colonization and deploy starbase mission on planning phase caused
override on mission already in progress. So AI could only colonizate or
deploy starbase if it had only one target. This kind of explains why
colonization was most effective on beginning of the game.
2. Most of the missions handling trekking with reroute method.
This was called almost everytime no matter if route was set. This caused
movement usage and therefore FTL never worked with AI.
3. makeRerouteBeforeFTLMoves used fleet.getMovesLeft in for loop
where doing the movement. Inside that for loop moves were also
decreased. This caused that regular speed 2 was actually speed 1 for
AI and speed 3 was actually speed 2.

Overall colonization and deploy starbase worked randomly and
most of the AI movement did not use FTL but regular 1 sector speed.