#209 Tune down Chiraloid race a bit.
Tune up a bit Centaurs and Teuthidaes.
Reason for this changes were that I run over 600 simulated games by AI
and Chiraloid seem to win way more often other races. Centaurs and
Teuthidaes got only couple of wins per 100 games. There fore I change
those two races so that they do not need extra food. So food
requirement is not set to 100%. This made those two a bit closer to Humans,
Greyans and Sporks. Chiraloids are either the strongest winner or second strongest.
They are quite close to the Mothoids or Mothoids are close to Chiraloids.