#240 Fixes human decling with war.
Previously this did not cause reputation lost, or news or activated defensive pacts.
Why? Because war was made very early in if clause and after that was check if war
is not yet on then do reputation, lost, news etc. This bug was done in 2017 when
fixing bug where multiple wars could be done against same realm.