#411 Adds diplomatic delegacy as mission for AI.
AI tries to send diplomatic delegacies if it has exhausted on war or
wants to win diplomatic or cultural. These can be send also if just
wants to be diplomatic. Improves also a bit of AI's winnning strategy
selection. This strategy is not that much focues yet.

Seems that in test games between AIs games either end by scoring or domination
and only one time diplomatic victory has occured.