#423 Changes Alliance government name to Union.
This is done since realms can be alliance. So it was possible to have
Alliance of Alliance of Terran and Alliance of Centaur which sounds really bad.

Changes also all government types to unique. Previously Alliance(Now union) was identical
to Democracy. Republic and Federation were identical. So were Guild and Enterprise. Kingdom and
Empire, and Horde and Clan. Now these have some minor differences.