#412 Adds own planet handling for Lithorians.
Fixes possible NPE. This NPE was added when food production was moved
into separate method of planet update one turn. It seems that
there is possibility to have starvation on planet and I am interested
how it can happen. AI should be able prevent starvation.
I could not reproduce it while trying it anymore, so I added some logging
that hopefully I can spot it later if it happens.