#430 Adds sound effects which are between 1-2 seconds to silence to 2s.
Seems that if SFX is shorter than 1 second it is not being played at
all. If it is longer than 1 second but less than 2. It will played about
1 second but not above. SFX longer than 2 seconds seems to work just
fine. This issue has been happened only with one machine. Did not
reproduce with another one even I update JDK from there.